+++ src/main/java/com/zika/chessbot/bot/TranspositionEntry.java
@@ -11,10 +11,10 @@ public class TranspositionEntry {
     private long key;
     private Flag flag;
     private Move move;
-    private double score;
+    private int score;
     private int depth;
 
-    public TranspositionEntry(long key, Flag flag, double score, int depth) {
+    public TranspositionEntry(long key, Flag flag, int score, int depth) {
         this.flag = flag;
         this.score = score;
         this.depth = depth;
